requirements.txt
openai
requests
Pillow
langchain
black
lorem
